safer data types multiplication

+++ src/rand.h
@@ -1,7 +1,7 @@
 #ifndef RAND_H
 #define RAND_H
 
-void set_seed(float new_seed);
+void set_seed(unsigned long new_seed);
 unsigned char rand_uchar(unsigned int max);
 unsigned short rand_ushort(unsigned int max);
